+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## 0.1.0 - 2017-04-19
 
+- Add `getCapabilities()` & `exception()`
 - Initialize library